0829 add llama2 统一测试

+++ pe/rerope_patch.py
@@ -100,8 +100,8 @@ def forward_with_rerope(
             )
         attn_weights = attn_weights + attention_mask
 
-    # upcast attention to fp32
-    attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=torch.float32).to(query_states.dtype)
+    # upcast attention to fp32 NOTE:, dtype=torch.float32
+    attn_weights = nn.functional.softmax(attn_weights, dim=-1).to(query_states.dtype)
     attn_output = torch.matmul(attn_weights, value_states)
 
     if attn_output.size() != (bsz, self.num_heads, q_len, self.head_dim):
